Add files via upload
The MNIST dataset is a popular dataset used for training and testing in machine learning for handwritten digit recognition. 
Structure of the MNIST dataset
The MNIST dataset is a collection of 70,000 handwritten digits (0-9), with each image being 28x28 pixels. Here is the dataset information in the specified format:

Number of Instances: 70,000 images
Number of Attributes: 784 (28x28 pixels)
Target: Column represents the digit (0-9) corresponding to the handwritten image
Pixel 1-784: Each pixel value (0-255) represents the grayscale intensity of the corresponding pixel in the image.
The dataset is divided into two main subsets:
Training Set: This consists of 60,000 images and their labels, commonly used for training machine learning models.
Test Set: Contains 10,000 images with their corresponding labels, used for evaluating the performance of trained models.